+++ server/main.py
@@ -12,7 +12,7 @@ from .api.v1.crud import download_quiz, generate_quiz, get_user_quiz_history
 from .app.db.routes import router as db_router
 from .app.db.core.connection import startUp
 from server.app.quiz.routers.quiz import router as quiz_router
-from .schemas.model.password_reset_model import PasswordResetRequest, PasswordResetResponse, RequestPasswordReset, MessageResponse
+# from ..schemas.model.password_reset_model import PasswordResetRequest, PasswordResetResponse, RequestPasswordReset, MessageResponse
 from .schemas.model import UserModel, LoginRequestModel, LoginResponseModel
 from .schemas.query import (
     GenerateQuizQuery,
@@ -44,6 +44,7 @@ from server.email_utils import send_otp_email
 import jwt
 from datetime import datetime, timedelta
 from .app.db.routes import router as db_router
+from .app.auth.routes import router as auth_router  
 
 app = FastAPI(lifespan=lifespan)
 
@@ -58,6 +59,7 @@ app.add_middleware(
 app.include_router(db_router)
 app.include_router(quiz_router, prefix="/api", tags=["quiz"])
 app.include_router(healthcheck.router, prefix="/api", tags=["healthcheck"])
+app.include_router(auth_router, prefix="/auth", tags=["authentication"])
 
 mock_db: List[UserModel] = []
 
@@ -68,153 +70,13 @@ def read_root():
 
 load_dotenv()  
 
-mock_db: list[UserModel] = []
-
-redis_client = Redis(host="localhost", port=6379, db=0, decode_responses=True)
-
-def generate_otp():
-    return str(random.randint(100000, 999999))  
-
-SECRET_KEY = os.getenv("JWT_SECRET_KEY")
-ALGORITHM = "HS256"
-
-def generate_verification_token(email: str):
-    expire = datetime.utcnow() + timedelta(minutes=5)  
-    payload = {"sub": email, "exp": expire}
-    return jwt.encode(payload, SECRET_KEY, algorithm=ALGORITHM)
-
-def decode_verification_token(token: str):
-    try:
-        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-        return payload["sub"]
-    except jwt.ExpiredSignatureError:
-        raise HTTPException(status_code=400, detail="Token expired")
-    except jwt.InvalidTokenError:
-        raise HTTPException(status_code=400, detail="Invalid token")
-
-@app.post("/register/")
-async def register_user(user: UserModel):
-    if any(u["email"] == user.email for u in mock_db):
-        raise HTTPException(status_code=400, detail="Email already registered")
-    
-    otp = generate_otp()
-    token = generate_verification_token(user.email)
-    
-    redis_client.setex(f"otp:{user.email}", 300, otp)  
-    redis_client.setex(f"token:{user.email}", 1800, token)  
-    
-    send_otp_task.delay(user.email, otp, token)
-    
-    
-    mock_db.append(user.dict())
-    
-    return {"message": "User registered. Please check your email for verification."}
-
-
-@app.post("/verify-otp/")
-async def verify_otp(email: str, otp: str):
-    stored_otp = redis_client.get(f"otp:{email}")
-    attempts = int(redis_client.get(f"attempts:{email}") or 0)
-    
-    if attempts >= 4:
-        raise HTTPException(status_code=403, detail="Too many attempts. Request a new OTP.")
-    if stored_otp is None:
-        raise HTTPException(status_code=400, detail="OTP expired or not requested.")
-    if otp != stored_otp:
-        redis_client.incr(f"attempts:{email}")  
-        raise HTTPException(status_code=401, detail="Invalid OTP. Try again.")
-    
-   
-    for user in mock_db:
-        if user["email"] == email:
-            user["is_verified"] = True
-            break
-    
-    
-    redis_client.delete(f"otp:{email}")
-    redis_client.delete(f"token:{email}")  
-    return {"message": "OTP verified successfully!"}
-
-@app.get("/verify-link/")
-async def verify_link(token: str):
-    try:
-        email = decode_verification_token(token)
-    except HTTPException as e:
-        raise e 
-    
-    
-    for user in mock_db:
-        if user["email"] == email:
-            user["is_verified"] = True
-            break
-    
-    redis_client.delete(f"token:{email}")
-    redis_client.delete(f"otp:{email}")  
-    return {"message": "Email verified successfully!"}
+@app.get("/")
+def root():
+    return {"message": "Welcome to Quiz Generator API"}
 
 @app.get("/users/", response_model=List[UserModel])
 def list_users():
     return mock_db
-
-@app.post("/login/", response_model=LoginResponseModel)
-def login(request: LoginRequestModel):
-    user = next(
-        (
-            u for u in mock_db 
-            if (u.username == request.username_or_email or u.email == request.username_or_email)
-            and u.password == request.password
-        ),
-        None
-    )
-    if user is None:
-        raise HTTPException(status_code=401, detail="Invalid credentials")
-    return {"message": "Login successful", "user": user}
-
-@app.post("/request-password-reset", response_model=MessageResponse)
-async def request_password_reset(request: RequestPasswordReset):
-    user = next((u for u in mock_db if u["email"] == request.email), None)
-    if not user:
-        return {"message": "If this email exists, reset instructions have been sent."}
-    
-    otp = generate_otp()
-    token = generate_verification_token(request.email)
-    
-    redis_client.setex(f"otp:{request.email}", 300, otp)
-    redis_client.setex(f"token:{request.email}", 1800, token)
-    
-    send_password_reset_email.delay(request.email, otp, token) 
-
-    return {"message": "If this email exists, reset instructions have been sent."}
-
-@app.post("/reset-password", response_model=PasswordResetResponse)
-async def reset_password(request: PasswordResetRequest):
-    user = next((u for u in mock_db if u["email"] == request.email), None)
-    if not user:
-        raise HTTPException(status_code=404, detail="User not found")
-
-    if request.reset_method == "otp":
-        stored_otp = redis_client.get(f"otp:{request.email}")
-        if stored_otp is None:
-            raise HTTPException(status_code=400, detail="OTP expired or not found")
-        if request.otp != stored_otp:
-            raise HTTPException(status_code=401, detail="Invalid OTP")
-    
-    elif request.reset_method == "token":
-        if not request.token:
-            raise HTTPException(status_code=400, detail="Token is required")
-        try:
-            email_from_token = decode_verification_token(request.token)
-        except HTTPException as e:
-            raise e
-        if email_from_token != request.email:
-            raise HTTPException(status_code=403, detail="Token email mismatch")
-    
-    user["password"] = request.new_password
-
-    redis_client.delete(f"otp:{request.email}")
-    redis_client.delete(f"token:{request.email}")
-
-    return PasswordResetResponse(message="Password reset successful", success=True)
 
 @app.post("/generate-quiz")
 async def generate_quiz_handler(query: GenerateQuizQuery = Body(...)) -> Dict[str, Any]:
