+++ server/schemas/model/__init__.py
@@ -2,3 +2,4 @@ from .login_request_model import LoginRequestModel
 from .login_response_model import LoginResponseModel
 from .user_model import UserModel
 from .quiz_questions_model import QuizQuestionsModel
+from .password_reset_model import PasswordResetRequest, PasswordResetResponse, RequestPasswordReset, ResetMethod, MessageResponse
